Mark GenerateDepTrees task as incompatible with configuration cache (#34)
The plugin is currently incompatible with Gradle projects using the configuration cache. This affects the JFrog VSCode and IntelliJ plugins as well as they use this plugin.

These changes mark the GenerateDepTrees task as incompatible with the Gradle configuration cache to prevent issues due to project access at execution time. This is a workaround - a better, more permanent solution would be to fix the issues causing configuration cache problems.